+++ src/InlineInput/NumberInlineInput.tsx
@@ -7,63 +7,72 @@
 import React, { FC } from 'react';
 
 import { isFactor } from '../Slider/factor';
-import { RangeProp } from '../Slider/rangeShape';
+import {
+    getStep,
+    isValues,
+    Range,
+    RangeOrValues,
+    useValidatedRangeOrValues,
+    Values,
+} from '../Slider/range';
 import InlineInput from './InlineInput';
 
 import './number-inline-input.scss';
 
-const isInRange = (
-    value: number,
-    { min, max, decimals = 0, step = 0, explicitRange = [] }: RangeProp
-) =>
+export interface Props {
+    disabled?: boolean;
+    value: number;
+    range: RangeOrValues;
+    onChange: (value: number) => void;
+    onChangeComplete?: (value: number) => void;
+}
+
+const isInValues = (value: number, values: Values) => values.includes(value);
+
+const isInRange = (value: number, { min, max, decimals, step }: Range) =>
     value >= min &&
     value <= max &&
     value === Number(value.toFixed(decimals)) &&
-    (step > 0 ? isFactor(value, step) : true) &&
-    (explicitRange.length > 0 ? explicitRange.indexOf(value) !== -1 : true);
+    (step == null ? true : isFactor(value - min, step));
 
-interface Props {
-    disabled?: boolean;
-    value: number;
-    range: RangeProp;
-    onChange: (number: number) => void;
-    onChangeComplete?: (number: number) => void;
-}
+const isValid = (value: number, rangeOrValues: RangeOrValues) =>
+    isValues(rangeOrValues)
+        ? isInValues(value, rangeOrValues)
+        : isInRange(value, rangeOrValues);
 
-const incrementValue = (
+const nextInValues = (
     current: number,
-    range: RangeProp,
+    values: Values,
+    steps: number
+): number | undefined => {
+    const currentIndex = values.indexOf(current);
+    const newIndex = currentIndex + steps;
+
+    return values[newIndex];
+};
+
+const nextInRange = (current: number, range: Range, steps: number) => {
+    const newValue = Number(
+        (current + steps * getStep(range)).toFixed(range.decimals)
+    );
+
+    if (newValue >= range.min && newValue <= range.max) {
+        return newValue;
+    }
+};
+
+const changeValueStepwise = (
+    current: number,
+    rangeOrValues: RangeOrValues,
     steps: number,
     action: (v: number) => void
 ) => {
-    if (steps === 0) return;
+    const nextValue = isValues(rangeOrValues)
+        ? nextInValues(current, rangeOrValues, steps)
+        : nextInRange(current, rangeOrValues, steps);
 
-    if (
-        typeof range.explicitRange !== 'undefined' &&
-        range.explicitRange.length > 0
-    ) {
-        const currentIndex = range.explicitRange.indexOf(current);
-        const newIndex = currentIndex + steps;
-
-        if (newIndex >= 0 && newIndex < range.explicitRange.length) {
-            action(range.explicitRange[newIndex]);
-        }
-
-        return;
-    }
-
-    const decimal =
-        range.decimals && typeof range.decimals !== 'undefined'
-            ? range.decimals
-            : 0;
-    const stepValue =
-        range.step && typeof range.step !== 'undefined'
-            ? range.step
-            : 0.1 ** decimal;
-    const newValue = Number((current + steps * stepValue).toFixed(decimal));
-
-    if (newValue >= range.min && newValue <= range.max) {
-        action(newValue);
+    if (nextValue != null) {
+        action(nextValue);
     }
 };
 
@@ -73,21 +82,25 @@ const NumberInlineInput: FC<Props> = ({
     range,
     onChange,
     onChangeComplete = () => {},
-}) => (
-    <InlineInput
-        className="number-inline-input"
-        disabled={disabled}
-        value={String(value)}
-        isValid={newValue => isInRange(Number(newValue), range)}
-        onChange={newValue => onChange(Number(newValue))}
-        onChangeComplete={newValue => onChangeComplete(Number(newValue))}
-        onKeyboardIncrementAction={() =>
-            incrementValue(value, range, 1, onChange)
-        }
-        onKeyboardDecrementAction={() =>
-            incrementValue(value, range, -1, onChange)
-        }
-    />
-);
+}) => {
+    useValidatedRangeOrValues(range);
+
+    return (
+        <InlineInput
+            className="number-inline-input"
+            disabled={disabled}
+            value={String(value)}
+            isValid={newValue => isValid(Number(newValue), range)}
+            onChange={newValue => onChange(Number(newValue))}
+            onChangeComplete={newValue => onChangeComplete(Number(newValue))}
+            onKeyboardIncrementAction={() =>
+                changeValueStepwise(value, range, 1, onChange)
+            }
+            onKeyboardDecrementAction={() =>
+                changeValueStepwise(value, range, -1, onChange)
+            }
+        />
+    );
+};
 
 export default NumberInlineInput;