+++ Changelog.md
@@ -6,6 +6,55 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+-   Specifying a range with explicit values for a `NumberInlineInput` or a
+    `Slider` was changed. When you before wrote an `NumberInlineInput` element
+    (same applies to a `Slider`) like this:
+
+```jsx
+<NumberInlineInput
+    values={[4]}
+    range={{ min: 3, max: 7, explicitRange: [3, 4, 6, 7] }}
+    onChange={[() => {}]}
+/>
+```
+
+Instead you now have to replace the whole range object with the array and write
+it like this:
+
+```jsx
+<NumberInlineInput
+    values={[4]}
+    range={[3, 4, 6, 7]}
+    onChange={[() => {}]}
+/>
+```
+
+It is of course still possible to specify a range without explicit values like
+this:
+
+```jsx
+<NumberInlineInput
+    values={[4]}
+    range={{ min: 3, max: 7 }}
+    onChange={[() => {}]}
+/>
+```
+
+### Fixed
+
+-   A `Slider` with ticks but without a defined step displayed the ticks
+    wrongly.
+
+### Steps to upgrade when using this package
+
+-   If you are using a `NumberInlineInput` or `Slider` with a range with the
+    `explicitRange` property, then you have to replace the whole range with the
+    array of the `explicitRange` property, as shown above.
+
 ## 6.12.1 - 2022-11-29
 
 ### Fixed